backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.24.0
langchain>=1.2.0
langchain-core>=1.2.0
langchain-google-genai>=4.2.0
langchain-community>=0.4.0
langchain-text-splitters>=1.1.0
google-generativeai==0.3.2
pymongo[srv]==4.6.1
motor==3.3.2
python-dotenv==1.0.0
pypdf==3.17.4
numpy>=1.24.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
rank-bm25>=0.2.2